+++ CHANGELOG.md
@@ -1,1 +1,11 @@
-## [1.0.0] - Created package
+## [1.0.0]
+
+* Created package
+
+## [1.1.0]
+
+* Fixed `maxLines`
+
+## [1.2.0]
+
+* fixed: `matchEntry` now is nullable because `matchEntries` can be empty